Small cleanups.
