+++ candle-nn/src/encoding.rs
@@ -1,61 +1,4 @@
 //! Encoding Utilities. (e.g., one-hot/cold encoding)
-//!
-//! # Examples
-//!
-//! ## One-hot encoding
-//!
-//! ```rust
-//! use candle::{Shape, Tensor, Device};
-//! use candle_nn::encoding::one_hot;
-//!
-//! let device = candle::Device::Cpu;
-//!
-//! let indices = Tensor::new(vec![vec![0f32, 2.], vec![1., -1.]], &device)?;
-//! let depth = 4;
-//!
-//! let allow_f64 = true;
-//! let on_value = 1.0;
-//! let off_value = 0.0;
-//!
-//! let one_hot = one_hot::<f32>(indices, depth, allow_f64, on_value, off_value)?;
-//!
-//! let expected_matrix = [
-//!     [[1.0, 0.0, 0.0, 0.0], [0.0, 0.0, 1.0, 0.0]],
-//!     [[0.0, 1.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0]],
-//! ];
-//!
-//! assert_eq!(one_hot.shape(), &Shape::from((2, 2, depth)));
-//!
-//! let matrix = one_hot.to_vec3::<f32>()?;
-//!
-//! assert_eq!(matrix, expected_matrix);
-//!```
-//! ## One-cold Encoding
-//!
-//! ```rust
-//! use candle::{Shape, Tensor, Device};
-//! use candle_nn::encoding::one_hot;
-//!
-//!
-//! let device = candle::Device::Cpu;
-//! let depth = 4;
-//! let indices = Tensor::new(vec![vec![0i64, 2], vec![1, 3]], &device)?;
-//!
-//! let allow_f64 = false;
-//! let on_value = 0;
-//! let off_value = 1;
-//!
-//! let one_cold = one_hot::<u8>(indices, depth, allow_f64, on_value, off_value)?;
-//!
-//! let expected_matrix = [[[0, 1, 1, 1], [1, 1, 0, 1]], [[1, 0, 1, 1], [1, 1, 1, 0]]];
-//!
-//! assert_eq!(one_cold.shape(), &Shape::from((2, 2, depth)));
-//!
-//! let matrix = one_cold.to_vec3::<u8>()?;
-//!
-//! assert_eq!(matrix, expected_matrix);
-//! ```
-//!
 
 use candle::{bail, DType, Result, Tensor, WithDType};
 
@@ -96,11 +39,7 @@ use candle::{bail, DType, Result, Tensor, WithDType};
 ///
 /// let indices = Tensor::new(vec![vec![0i64, 2], vec![1, -1]], &device).unwrap();
 /// let depth = 4;
-///
-/// let on_value = 1.0;
-/// let off_value = 0.0;
-///
-/// let one_hot = one_hot::<f32>(indices, depth, on_value, off_value).unwrap();
+/// let one_hot = one_hot(indices, depth, 1f32, 0f32).unwrap();
 ///
 /// let expected_matrix = [
 ///     [[1.0, 0.0, 0.0, 0.0], [0.0, 0.0, 1.0, 0.0]],
@@ -123,12 +62,7 @@ use candle::{bail, DType, Result, Tensor, WithDType};
 /// let device = candle::Device::Cpu;
 /// let depth = 4;
 /// let indices = Tensor::new(vec![vec![0u8, 2], vec![1, 3]], &device).unwrap();
-///
-/// let on_value: u8 = 0;
-/// let off_value = 1;
-///
-/// // Note: the method does not require the turbofish operator, as the type is inferred from the `on_value` and `off_value`.
-/// let one_cold = one_hot(indices, depth, on_value, off_value).unwrap();
+/// let one_cold = one_hot(indices, depth, 0u8, 1u8).unwrap();
 ///
 /// let expected_matrix = [[[0, 1, 1, 1], [1, 1, 0, 1]], [[1, 0, 1, 1], [1, 1, 1, 0]]];
 ///
@@ -142,13 +76,11 @@ use candle::{bail, DType, Result, Tensor, WithDType};
 ///
 /// # Bails
 ///
-/// This method will bail on tensors with a rank greater than 3.
-///
-/// This method will bail if the index value is less than -1.
-///
-/// This method will bail if the index value is greater than or equal to the depth value.
-///
-/// This method will bail if the input data type is not one of `U8`, `U32` or `I64`.
+/// This method bails if:
+/// - The input tensor has a rank greater than 3.
+/// - One of the index value is less than -1.
+/// - One of the index value is greater than or equal to the depth value.
+/// - The input data type is not `U8`, `U32`, or `I64`.
 ///
 /// # API Design
 ///
@@ -165,24 +97,22 @@ pub fn one_hot<D: WithDType>(
     match rank {
         0 => {
             let mut v = vec![off_value; depth];
-            let idx = 0;
             match dtype {
                 DType::U8 => {
                     let vi = indices.to_vec0::<u8>()?;
-                    set_usize_value(vi as usize, idx, depth, &mut v, on_value)?;
+                    set_usize_value(vi as usize, 0, depth, &mut v, on_value)?;
                 }
                 DType::U32 => {
                     let vi = indices.to_vec0::<u32>()?;
-                    set_usize_value(vi as usize, idx, depth, &mut v, on_value)?;
+                    set_usize_value(vi as usize, 0, depth, &mut v, on_value)?;
                 }
                 DType::I64 => {
                     let vi = indices.to_vec0::<i64>()?;
-                    set_int64_value(vi, idx, depth, &mut v, on_value)?;
+                    set_int64_value(vi, 0, depth, &mut v, on_value)?;
                 }
                 d => unsupported_dtype(d)?,
             };
-
-            Tensor::new(v, indices.device())
+            Tensor::from_vec(v, (depth,), indices.device())
         }
         1 => {
             let dim1 = indices.dims1()?;
@@ -190,38 +120,34 @@ pub fn one_hot<D: WithDType>(
 
             match dtype {
                 DType::U8 => {
-                    let index = indices.to_vec1::<u8>()?;
-                    for i in 0..dim1 {
-                        set_usize_value(index[i] as usize, i * depth, depth, &mut v, on_value)?;
+                    let indices = indices.to_vec1::<i64>()?;
+                    for (i, &index) in indices.iter().enumerate() {
+                        set_usize_value(index as usize, i * depth, depth, &mut v, on_value)?;
                     }
                 }
                 DType::U32 => {
-                    let index = indices.to_vec1::<u32>()?;
-                    for i in 0..dim1 {
-                        set_usize_value(index[i] as usize, i * depth, depth, &mut v, on_value)?;
+                    let indices = indices.to_vec1::<i64>()?;
+                    for (i, &index) in indices.iter().enumerate() {
+                        set_usize_value(index as usize, i * depth, depth, &mut v, on_value)?;
                     }
                 }
                 DType::I64 => {
-                    let index = indices.to_vec1::<i64>()?;
-                    for i in 0..dim1 {
-                        set_int64_value(index[i], i * depth, depth, &mut v, on_value)?;
+                    let indices = indices.to_vec1::<i64>()?;
+                    for (i, &index) in indices.iter().enumerate() {
+                        set_int64_value(index, i * depth, depth, &mut v, on_value)?;
                     }
                 }
                 d => unsupported_dtype(d)?,
             };
-
-            Tensor::new(v, indices.device())?.reshape(&[dim1, depth])
+            Tensor::from_vec(v, (dim1, depth), indices.device())
         }
         2 => {
             let (dim1, dim2) = indices.dims2()?;
             let mut v = vec![off_value; depth * dim1 * dim2];
-
             let idx = |i: usize, j: usize, depth: usize, dim2: usize| -> usize {
                 i * depth * dim2 + j * depth
             };
-
             let iter = (0..dim1).flat_map(|i| (0..dim2).map(move |j| (i, j)));
-
             match dtype {
                 DType::U8 => {
                     let index = indices.to_vec2::<u8>()?;
@@ -261,21 +187,17 @@ pub fn one_hot<D: WithDType>(
                 }
                 d => unsupported_dtype(d)?,
             };
-
-            Tensor::new(v, indices.device())?.reshape(&[dim1, dim2, depth])
+            Tensor::from_vec(v, (dim1, dim2, depth), indices.device())
         }
         3 => {
             let (dim1, dim2, dim3) = indices.dims3()?;
             let mut v = vec![off_value; depth * dim1 * dim2 * dim3];
-
             let idx =
                 |i: usize, j: usize, k: usize, depth: usize, dim2: usize, dim3: usize| -> usize {
                     i * depth * dim2 * dim3 + j * depth * dim3 + k * depth
                 };
-
             let iter = (0..dim1)
                 .flat_map(|i| (0..dim2).flat_map(move |j| (0..dim3).map(move |k| (i, j, k))));
-
             match dtype {
                 DType::U8 => {
                     let index = indices.to_vec3::<u8>()?;
@@ -315,8 +237,7 @@ pub fn one_hot<D: WithDType>(
                 }
                 d => unsupported_dtype(d)?,
             };
-
-            Tensor::new(v, indices.device())?.reshape(&[dim1, dim2, dim3, depth])
+            Tensor::from_vec(v, (dim1, dim2, dim3, depth), indices.device())
         }
         _ => {
             bail!("one_hot: rank {} is not supported.", rank)
@@ -324,46 +245,8 @@ pub fn one_hot<D: WithDType>(
     }
 }
 
-// Bail if the value is less than -1.
-fn check_negative(value: i64) -> Result<()> {
-    if value < -1 {
-        bail!("one_hot: Invalid negative index value. Expected a positive index value or `-1` value to ignore. Received {}", value);
-    }
-    Ok(())
-}
-
-// Bail if the value is greater than the depth.
-fn check_value_bounds(value: usize, depth: usize) -> Result<()> {
-    if value >= depth {
-        bail!(
-            "one_hot: Index value exceeds the depth value. Index value, {}, must be less than the depth value, {}.",
-            value,
-            depth
-        )
-    }
-
-    Ok(())
-}
-
-// Bail if the index is out of bounds.
-fn check_idx(idx: usize, len: usize) -> Result<()> {
-    if idx >= len {
-        bail!(
-            "one_hot: Index out of bounds. Expected index value, {}, to be less than {}.",
-            idx,
-            len
-        );
-    }
-
-    Ok(())
-}
-
-// Bail if the data type is not supported.
 fn unsupported_dtype(dtype: DType) -> Result<()> {
-    bail!(
-        "one_hot: Unsupported data type. Expected U8, U32, or I64. Received {:?}",
-        dtype
-    );
+    bail!("one_hot: unsupported data type {dtype:?}, expected U8, U32, or I64")
 }
 
 // Set unsigned usize index values to the given value.
@@ -374,13 +257,14 @@ fn set_usize_value<D: WithDType>(
     v: &mut Vec<D>,
     on_value: D,
 ) -> Result<()> {
-    check_value_bounds(value, depth)?;
-
+    if value >= depth {
+        bail!("one_hot: index value {value} exceeds depth {depth}")
+    }
     let idx = idx + value;
-
-    check_idx(idx, v.len())?;
+    if idx >= v.len() {
+        bail!("one_hot: index out of bounds {idx}, len {}", v.len());
+    }
     v[idx] = on_value;
-
     Ok(())
 }
 
@@ -400,137 +284,10 @@ fn set_int64_value<D: WithDType>(
     if value == -1 {
         return Ok(());
     }
-
-    // Bail if the index value is less than -1
-    check_negative(value)?;
-
-    // Since i64 cannot be negative after the previous check,
-    // we can safely cast to usize.
+    if value < -1 {
+        bail!(
+            "one_hot: invalid negative index value {value}, expected a positive index value or -1"
+        );
+    }
     set_usize_value(value as usize, idx, depth, v, on_value)
-}
-
-#[cfg(test)]
-mod tests {
-    use candle::Shape;
-
-    use super::*;
-
-    #[test]
-    fn test_i64_one_hot() -> std::result::Result<(), Box<dyn std::error::Error>> {
-        let device = candle::Device::Cpu;
-
-        let indices = Tensor::new(vec![vec![0i64, 2], vec![1, -1]], &device)?;
-        let depth = 4;
-
-        let on_value = 1.0;
-        let off_value = 0.0;
-
-        let one_hot = one_hot::<f32>(indices, depth, on_value, off_value)?;
-
-        let expected_matrix = [
-            [[1., 0., 0., 0.], [0., 0., 1., 0.]],
-            [[0., 1., 0., 0.], [0., 0., 0., 0.]],
-        ];
-
-        assert_eq!(one_hot.shape(), &Shape::from((2, 2, depth)));
-
-        let matrix = one_hot.to_vec3::<f32>()?;
-
-        assert_eq!(matrix, expected_matrix);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_rank_3_one_hot() -> std::result::Result<(), Box<dyn std::error::Error>> {
-        let device = candle::Device::Cpu;
-
-        let indices = Tensor::new(
-            vec![
-                vec![vec![0i64, 1], vec![2, 3]],
-                vec![vec![3, 1], vec![1, -1]],
-            ],
-            &device,
-        )?;
-        let depth = 4;
-
-        let on_value = 1.0;
-        let off_value = 0.0;
-
-        let one_hot = one_hot::<f32>(indices, depth, on_value, off_value)?;
-
-        let expected_matrix = Tensor::new(
-            vec![
-                vec![
-                    vec![vec![1f32, 0., 0., 0.], vec![0., 1., 0., 0.]],
-                    vec![vec![0., 0., 1., 0.], vec![0., 0., 0., 1.]],
-                ],
-                vec![
-                    vec![vec![0., 0., 0., 1.], vec![0., 1., 0., 0.]],
-                    vec![vec![0., 1., 0., 0.], vec![0., 0., 0., 0.]],
-                ],
-            ],
-            &device,
-        )?;
-
-        assert_eq!(one_hot.shape(), expected_matrix.shape());
-        assert_eq!(one_hot.dims(), expected_matrix.dims());
-
-        let matrix = one_hot.get(1)?.to_vec3::<f32>()?;
-        let expected_matrix = expected_matrix.get(1)?.to_vec3::<f32>()?;
-
-        assert_eq!(matrix, expected_matrix);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_u8_one_cold() -> std::result::Result<(), Box<dyn std::error::Error>> {
-        let device = candle::Device::Cpu;
-        let depth = 4;
-        let indices = Tensor::new(vec![vec![0i64, 2], vec![1, -1]], &device)?;
-
-        let on_value = 0u8;
-        let off_value = 1;
-
-        // Note that the method does not require the turbofish operator, as the type is inferred from the on_value.
-        let one_cold = one_hot(indices, depth, on_value, off_value)?;
-
-        let expected_matrix = [[[0, 1, 1, 1], [1, 1, 0, 1]], [[1, 0, 1, 1], [1, 1, 1, 1]]];
-
-        assert_eq!(one_cold.shape(), &Shape::from((2, 2, depth)));
-
-        let matrix = one_cold.to_vec3::<u8>()?;
-
-        assert_eq!(matrix, expected_matrix);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_iter() -> std::result::Result<(), Box<dyn std::error::Error>> {
-        let device = candle::Device::Cpu;
-        let depth = 4;
-        let indices = Tensor::new(vec![vec![0i64, 2], vec![1, -1]], &device)?;
-        let matrix = indices.to_vec2::<i64>()?;
-        let (dim1, dim2) = indices.dims2()?;
-
-        let iter = (0..dim1).flat_map(|i| (0..dim2).map(move |j| (i, j)));
-
-        let mut v = vec![0; depth * dim1 * dim2];
-
-        for (i, j) in iter {
-            let idx = i * depth * dim2 + j * depth;
-            v[idx] = matrix[i][j];
-        }
-
-        for i in 0..dim1 {
-            for j in 0..dim2 {
-                let idx = i * depth * dim2 + j * depth;
-                assert_eq!(v[idx], matrix[i][j]);
-            }
-        }
-
-        Ok(())
-    }
 }
